feat(server): add function recycle bin

+++ server/src/function/function.controller.ts
@@ -10,12 +10,14 @@ import {
   HttpException,
   HttpStatus,
   Req,
+  Query,
 } from '@nestjs/common'
 import { CreateFunctionDto } from './dto/create-function.dto'
 import { UpdateFunctionDto } from './dto/update-function.dto'
 import {
   ApiResponseArray,
   ApiResponseObject,
+  ApiResponsePagination,
   ResponseUtil,
 } from '../utils/response'
 import { ApiBearerAuth, ApiOperation, ApiTags } from '@nestjs/swagger'
@@ -30,6 +32,8 @@ import { ApplicationAuthGuard } from 'src/authentication/application.auth.guard'
 import { CloudFunctionHistory } from './entities/cloud-function-history'
 import { CloudFunction } from './entities/cloud-function'
 import { UpdateFunctionDebugDto } from './dto/update-function-debug.dto'
+import { CloudFunctionRecycleBin } from './entities/cloud-function-recycle-bin'
+import { CloudFunctionRecycleBinQuery } from './interface/cloud-function-query.interface'
 
 @ApiTags('Function')
 @ApiBearerAuth('Authorization')
@@ -214,7 +218,12 @@ export class FunctionController {
         HttpStatus.NOT_FOUND,
       )
     }
-
+    const recycleBinStorage = await this.functionsService.getRecycleBinStorage(
+      appid,
+    )
+    if (recycleBinStorage >= 1000) {
+      return ResponseUtil.error('Recycle bin is full, please free up space')
+    }
     const res = await this.functionsService.removeOne(func)
     if (!res) {
       return ResponseUtil.error(i18n.t('function.delete.error'))
@@ -276,5 +285,33 @@ export class FunctionController {
 
     const res = await this.functionsService.getHistory(func)
     return ResponseUtil.ok(res)
+  }
+
+  /**
+   * Get function history
+   */
+  @ApiOperation({ summary: 'Get cloud function recycle bin' })
+  @ApiResponsePagination(CloudFunctionRecycleBin)
+  @UseGuards(JwtAuthGuard, ApplicationAuthGuard)
+  @Get('recycle-bin')
+  async getRecycleBin(
+    @Param('appid') appid: string,
+    @Query('keyword') keyword?: string,
+    @Query('page') page?: number,
+    @Query('pageSize') pageSize?: number,
+    @Query('startTime') startTime?: number,
+    @Query('endTime') endTime?: number,
+  ) {
+    const query: CloudFunctionRecycleBinQuery = {
+      page: page || 1,
+      pageSize: pageSize || 12,
+    }
+    if (query.pageSize > 100) {
+      query.pageSize = 100
+    }
+    if (keyword) {
+      query.name = keyword
+    }
+    // return ResponseUtil.ok(res)
   }
 }
